custom HTTP capabilities moved to their own subpackage
