+++ basculehttp/testSuite_test.go
@@ -62,3 +62,21 @@ func (suite *TestSuite) newAuthorizationParser(opts ...AuthorizationParserOption
 	suite.Require().NotNil(ap)
 	return ap
 }
+
+// newAuthenticator creates a bascule.Authenticator that is expected to be valid.
+// Assertions as to validity are made prior to returning.
+func (suite *MiddlewareTestSuite) newAuthenticator(opts ...bascule.AuthenticatorOption[*http.Request]) *bascule.Authenticator[*http.Request] {
+	a, err := NewAuthenticator(opts...)
+	suite.Require().NoError(err)
+	suite.Require().NotNil(a)
+	return a
+}
+
+// newAuthorizer creates a bascule.Authorizer that is expected to be valid.
+// Assertions as to validity are made prior to returning.
+func (suite *MiddlewareTestSuite) newAuthorizer(opts ...bascule.AuthorizerOption[*http.Request]) *bascule.Authorizer[*http.Request] {
+	a, err := NewAuthorizer(opts...)
+	suite.Require().NoError(err)
+	suite.Require().NotNil(a)
+	return a
+}
